requirements.txt
uvicorn==0.18.3
fastapi==0.80.0
Pillow==9.2.0
pydantic==1.9.2
pymongo==4.2.0
python_Levenshtein==0.12.2
requests==2.28.1
sentence_transformers==2.2.2
torch==1.12.1
torchvision==0.13.1
